requirements.txt
aiogram==2.9.2
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
python-dotenv==1.0.1
requests==2.31.0
stripe==8.1.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
